+++ src/components/ModalDeletePieces/index.js
@@ -1,0 +1,3 @@
+import ModalDeletePiecesContainer from './ModalDeletePiecesContainer';
+
+export { ModalDeletePiecesContainer as default };