+++ src/components/Utils/order/showUpdateOrderError.js
@@ -21,7 +21,13 @@ const showMessage = (callout, code, error, path) => {
   });
 };
 
-const showUpdateOrderError = async (response, callout, openModal, genericCode = 'orderGenericError1') => {
+const showUpdateOrderError = async (
+  response,
+  callout,
+  openModal,
+  genericCode = 'orderGenericError1',
+  toggleDeletePieces = null,
+) => {
   let error;
 
   try {
@@ -49,6 +55,10 @@ const showUpdateOrderError = async (response, callout, openModal, genericCode = 
 
       errors = errors.length ? errors : [{ code }];
       openModal(errors);
+      break;
+    }
+    case ERROR_CODES.piecesNeedToBeDeleted: {
+      if (toggleDeletePieces) toggleDeletePieces();
       break;
     }
     case ERROR_CODES.missingInstanceStatus: {