+++ test/bigtest/interactors/index.js
@@ -1,0 +1,1 @@
+export { default as DeletePiecesModal } from './delete-pieces-modal';