+++ test/bigtest/interactors/order-lines.js
@@ -36,7 +36,7 @@ import { TIMEOUT } from './const';
 export default interactor(class OrderLinesInteractor {
   static defaultScope = '[data-test-order-line-instances]';
 
-  instances = collection('[role=group] [role=row]');
+  instances = collection('[data-row-inner]');
 
   navigation = new OrdersNavigation();